old/requirements.txt
# Odin's Eye — CV Term Project
# Requirements for Apple Silicon (M4 Max) + Colab (A100/H100)

# ── Core ──
ultralytics>=8.4.0
torch>=2.1.0
torchvision>=0.16.0
Pillow
tqdm
numpy
matplotlib
opencv-python
scipy

# ── Phase 0: Calibration ──
shapely
networkx

# ── Phase 1 v1: YOLO + BoT-SORT + OSNet ──
gdown

# ── Phase 1 v2: SAM2.1 + D4SM + DINOv2 ──
# sam2 is installed from d4sm fork via setup_tracking_v2.sh
hydra-core
omegaconf

# ── Phase 3: VLM Narration ──
mlx
mlx-vlm
huggingface_hub

# ── Dataset prep ──
roboflow
kagglehub